requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.8
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
ipdb==0.12.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Markdown==3.1.1
oauthlib==3.1.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.7
